+++ SecTester.sln
@@ -84,8 +84,8 @@ Global
 		{F5E1A845-BDCF-46D0-A8FE-CE47D41D35F7} = {27E70918-C7EB-4711-BCB9-CE1C9C0A5C5B}
 		{A48670DA-9103-4AAB-8914-6B7FDCD8654E} = {27E70918-C7EB-4711-BCB9-CE1C9C0A5C5B}
 		{64E65289-A36D-46C6-9410-6B2040A51E8A} = {27E70918-C7EB-4711-BCB9-CE1C9C0A5C5B}
-				{5608AF78-4476-47F0-8CEB-FD9849E8778B} = {2C77C898-B581-47F6-A8BF-3C1EE589F052}
-				{0B9D9422-FBB0-4B21-B3C1-9927DDEDA702} = {2C77C898-B581-47F6-A8BF-3C1EE589F052}
+		{5608AF78-4476-47F0-8CEB-FD9849E8778B} = {2C77C898-B581-47F6-A8BF-3C1EE589F052}
+		{0B9D9422-FBB0-4B21-B3C1-9927DDEDA702} = {2C77C898-B581-47F6-A8BF-3C1EE589F052}
 		{853E5140-4448-41EA-8F2B-9C9FF2B5491E} = {2C77C898-B581-47F6-A8BF-3C1EE589F052}
 		{368F1410-F710-4678-8376-7FC356CFD338} = {27E70918-C7EB-4711-BCB9-CE1C9C0A5C5B}
 		{5FC5E230-64F6-45F1-8661-1FF856FA6BC6} = {2C77C898-B581-47F6-A8BF-3C1EE589F052}
