+++ src/SecTester.Core/RetryStrategies/ExponentialBackoffOptions.cs
@@ -1,3 +1,3 @@
-namespace SecTester.Repeater.RetryStrategies;
+namespace SecTester.Core.RetryStrategies;
 
 public record ExponentialBackoffOptions(int MaxDepth = 3, int MinInterval = 50);
