+++ src/SecTester.Core/SecTester.Core.csproj
@@ -6,9 +6,15 @@
 
   <ItemGroup>
     <Folder Include="Bus"/>
+    <Folder Include="Commands"/>
     <Folder Include="CredentialProviders"/>
+    <Folder Include="Dispatchers"/>
+    <Folder Include="Exceptions"/>
     <Folder Include="Extensions"/>
     <Folder Include="Internal"/>
+    <Folder Include="Logger"/>
+    <Folder Include="RetryStrategies"/>
+    <Folder Include="Utils"/>
   </ItemGroup>
 
   <ItemGroup>
@@ -16,7 +22,23 @@
   </ItemGroup>
 
   <ItemGroup>
+    <PackageReference Include="System.Linq.Async" Version="6.0.1"/>
+    <PackageReference Include="System.Text.Json" Version="7.0.3"/>
+    <PackageReference Include="System.Threading.RateLimiting" Version="7.0.0"/>
+    <PackageReference Include="Microsoft.Bcl.AsyncInterfaces" Version="7.0.0"/>
+    <PackageReference Include="Microsoft.Extensions.Http" Version="6.0.0"/>
     <PackageReference Include="Microsoft.Extensions.Logging" Version="6.0.0"/>
     <PackageReference Include="Microsoft.Extensions.Logging.Console" Version="6.0.0"/>
+    <PackageReference Include="Macross.Json.Extensions" Version="3.0.0"/>
+  </ItemGroup>
+
+  <ItemGroup>
+    <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleTo">
+      <_Parameter1>$(MSBuildProjectName).Tests</_Parameter1>
+    </AssemblyAttribute>
+    <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleTo">
+      <_Parameter1>DynamicProxyGenAssembly2</_Parameter1>
+    </AssemblyAttribute>
   </ItemGroup>
 </Project>
+
