+++ src/SecTester.Repeater/Api/DeleteRepeaterRequest.cs
@@ -1,6 +1,6 @@
 using System.Net.Http;
 using SecTester.Core;
-using SecTester.Repeater.Commands;
+using SecTester.Core.Commands;
 
 namespace SecTester.Repeater.Api;
