+++ src/SecTester.Scan/Commands/GetScan.cs
@@ -1,4 +1,4 @@
-using SecTester.Repeater.Commands;
+using SecTester.Core.Commands;
 using SecTester.Scan.Models;
 
 namespace SecTester.Scan.Commands;
