+++ src/SecTester.Scan/Commands/ListIssues.cs
@@ -1,5 +1,5 @@
 using System.Collections.Generic;
-using SecTester.Repeater.Commands;
+using SecTester.Core.Commands;
 using SecTester.Scan.Models;
 
 namespace SecTester.Scan.Commands;
