+++ src/SecTester.Scan/Commands/UploadHar.cs
@@ -1,8 +1,8 @@
 using System.Collections.Generic;
 using System.Net.Http;
 using System.Text;
-using SecTester.Repeater.Commands;
-using SecTester.Repeater.Dispatchers;
+using SecTester.Core.Commands;
+using SecTester.Core.Dispatchers;
 using SecTester.Scan.Models;
 
 namespace SecTester.Scan.Commands;
