+++ test/SecTester.Core.Tests/Fixtures/FooBaz.cs
@@ -1,3 +1,3 @@
-namespace SecTester.Repeater.Tests.Fixtures;
+namespace SecTester.Core.Tests.Fixtures;
 
 internal record FooBaz(IEnumerable<KeyValuePair<string, IEnumerable<string>>> Headers);
