+++ test/SecTester.Core.Tests/Fixtures/IMockInterface.cs
@@ -1,4 +1,4 @@
-namespace SecTester.Repeater.Tests.Fixtures;
+namespace SecTester.Core.Tests.Fixtures;
 
 internal interface IMockInterface
 {
