+++ test/SecTester.Core.Tests/Usings.cs
@@ -1,6 +1,9 @@
 global using System;
 global using System.Collections.Generic;
 global using System.Globalization;
+global using System.Text;
+global using System.Text.Json;
+global using System.Text.Json.Serialization;
 global using FluentAssertions;
 global using FluentAssertions.Extensions;
 global using Microsoft.Extensions.DependencyInjection;
@@ -11,9 +14,15 @@ global using Microsoft.Extensions.Options;
 global using NSubstitute;
 global using NSubstitute.ClearExtensions;
 global using NSubstitute.ExceptionExtensions;
+global using RichardSzalay.MockHttp;
 global using SecTester.Core.Bus;
+global using SecTester.Core.Commands;
 global using SecTester.Core.CredentialProviders;
+global using SecTester.Core.Dispatchers;
+global using SecTester.Core.Exceptions;
 global using SecTester.Core.Extensions;
+global using SecTester.Core.RetryStrategies;
+global using SecTester.Core.Exceptions;
 global using SecTester.Core.Logger;
 global using SecTester.Core.Utils;
 global using Xunit;
