--- test/SecTester.Repeater.Tests/Fixtures/ConcreteCommand.cs
@@ -1,8 +1,0 @@
-namespace SecTester.Repeater.Tests.Fixtures;
-
-internal record ConcreteCommand : Command<Unit>
-{
-  public string Payload { get; }
-
-  public ConcreteCommand(string payload, bool? expectReply = null, TimeSpan? ttl = null) : base(expectReply, ttl) => Payload = payload;
-}
