+++ test/SecTester.Repeater.Tests/Usings.cs
@@ -17,16 +17,15 @@ global using NSubstitute.ReturnsExtensions;
 global using RichardSzalay.MockHttp;
 global using SecTester.Core;
 global using SecTester.Core.Bus;
+global using SecTester.Core.Commands;
+global using SecTester.Core.Dispatchers;
 global using SecTester.Core.Exceptions;
+global using SecTester.Core.RetryStrategies;
 global using SecTester.Core.Logger;
 global using SecTester.Core.Utils;
 global using SecTester.Repeater.Api;
 global using SecTester.Repeater.Bus;
-global using SecTester.Repeater.Commands;
-global using SecTester.Repeater.Dispatchers;
-global using SecTester.Repeater.Exceptions;
 global using SecTester.Repeater.Extensions;
-global using SecTester.Repeater.RetryStrategies;
 global using SecTester.Repeater.Runners;
 global using SecTester.Repeater.Tests.Mocks;
 global using Xunit;
