+++ test/SecTester.Scan.Tests/Usings.cs
@@ -14,8 +14,8 @@ global using SecTester.Core;
 global using SecTester.Core.Bus;
 global using SecTester.Core.Exceptions;
 global using SecTester.Core.Utils;
-global using SecTester.Repeater.Dispatchers;
-global using SecTester.Repeater.Extensions;
+global using SecTester.Core.Dispatchers;
+global using SecTester.Core.Extensions;
 global using SecTester.Scan.CI;
 global using SecTester.Scan.Commands;
 global using SecTester.Scan.Exceptions;
